% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbanker.R
\name{adbanker}
\alias{adbanker}
\title{A function for adbanking a single vector.}
\usage{
adbanker(data, adbank_values = adbanks_default)
}
\arguments{
\item{data}{the data.frame that will be used in the adbank transformation ("is adbanked"). Make sure to use proper and correct variable names!}
}
\description{
This function will compute a range of carry-over depot effects for testing in a linear MMM model.
}
\examples{
adbanker(mtcars[,c("wt", "qsec")])
}
\keyword{ADBANKs}
\keyword{a}
\keyword{compute}
\keyword{data.frame}
\keyword{for}
