% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbankr.R
\name{adbankr}
\alias{adbankr}
\title{A function for adbanking a single vector.}
\usage{
adbankr(data_vec)
}
\arguments{
\item{data_vec}{the vector that will be adbanked. Can iterate over any number of variables in a dataframe using purr::map.}
}
\description{
This function will compute a range of carry-over depot effects for testing in a linear MMM model.
}
\examples{
adbankr()
}
\keyword{ADBANKs}
\keyword{a}
\keyword{compute}
\keyword{for}
\keyword{vector}
