% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternate_decomp.R
\name{alternate_decomp}
\alias{alternate_decomp}
\title{Computes alternative and somewhat simpler version of the KPI decomposition}
\usage{
alternate_decomp(model = model1)
}
\arguments{
\item{model}{Name of the model object that will be decomped. Defaults to model1.}
}
\description{
This function decomps a model in a very basic and intuitive manner: The coefficients are simply multiplied with the respective data column, and shares are computed on the sum of these columns. Negative effects are computed in the same step.
}
\examples{
alternate_decomp()
}
\keyword{KPI}
\keyword{alternate,}
\keyword{decomposition.}
\keyword{model}
\keyword{straightforward}
