% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{finetune}
\alias{finetune}
\title{A function for finetuning the effects of media carry-over}
\usage{
finetune(model = model1, var_name, data = full_data)
}
\arguments{
\item{model}{Input model. Defaults to model1.}

\item{var_name}{Part of the variable name, such as TV_GRP. Can be used like a wildcard.}

\item{data}{The dataset to search for variables. Defaults to full_data.}
}
\description{
This function allows you a quick printout of all relevant effects for a media group
}
\examples{
finetune(my_model, "radio_spend", my_data) ## will search for all variables that contain the string "radio_spend"
pp()
}
\keyword{default}
\keyword{more}
\keyword{print}
\keyword{than}
\keyword{the}
