% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtally.R
\name{gtally}
\alias{gtally}
\title{A function for fast grouping and tallying of variables}
\usage{
gtally(data, group)
}
\arguments{
\item{data}{Input data. Intended use for this function is in a tidyverse pipe, so you shouldn't have to specify this argument, usually}

\item{group}{which variable to group and tally by}
}
\description{
This function allows you a quick printout of all that's relevant to you
}
\examples{
pp(15) ## this will print 15 lines instead of the tidyverse standard of 10
pp()
}
\keyword{default}
\keyword{more}
\keyword{print}
\keyword{than}
\keyword{the}
