% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout.R
\name{holdout}
\alias{holdout}
\title{A function for computing a holdout test on the model file}
\usage{
holdout(model = model1)
}
\arguments{
\item{model}{Input data. Just name the model. Defaults to "model1".}
}
\description{
This function allows you to easily calculate a time-series holdout test.
}
\examples{
preview_torte(my_default_model)
preview_torte()
}
\keyword{chart.}
\keyword{decomposition}
\keyword{pie}
\keyword{preview}
\keyword{quick}
\keyword{sales}
