% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korrelierer.R
\name{korrelierer}
\alias{korrelierer}
\title{A function for finding correlated VARs by name}
\usage{
korrelierer(daten, modell = model1, variable = "TV")
}
\arguments{
\item{daten}{the data that you wish to find CORed VARs in. Will be searched for any VAR containing the variable string}

\item{modell}{name the model that your trying to find CORs for (that's where the FUN get the RESIDs) - defaults to model1.}

\item{variable}{name the VAR you wish to find CORs for. defaults to "TV"}
}
\description{
This function searches for correlations towards the residuals and ranks them by strength..
}
\examples{
korrelierer()
}
\keyword{CORs}
\keyword{NAME}
\keyword{by}
\keyword{find}
