% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagger.R
\name{lagger}
\alias{lagger}
\title{Functions to generate lagged and logged variables}
\usage{
lagger(x, lag = 1)
}
\arguments{
\item{x}{whatever data you put into it. works with standard DFs from HexView.
very general, may throw NAs. typically these do not pose a problem for further modelling.
Function will be updated in the near future.}
}
\description{
These functions transforms input data to natural logarithm form needed for a diminishing returns media analysis (logger)
or lagged variables. It is recommended to generated logged forms first and then apply the lagger() function over the netire dataset.
}
\examples{
lagger(mtcars)
logger()
}
\keyword{log}
\keyword{transform}
