% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl.R
\name{mdl}
\alias{mdl}
\title{A function for adbanking a single vector.}
\usage{
mdl(..., data = full_data)
}
\arguments{
\item{...}{the formula for base R's lm() function.}

\item{data}{data containing the variables in the formula. Defaults to full_data.}
}
\description{
This function will compute a range of carry-over depot effects for testing in a linear MMM model.
}
\examples{
mdl()
}
\keyword{MMM}
\keyword{estimate}
\keyword{model}
