% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_painter.R
\name{model_painter}
\alias{model_painter}
\title{A function for painting all these happy little observations and predictions found in your MMM model}
\usage{
model_painter(modell = model1)
}
\arguments{
\item{modell}{the model you have created. Dependent variable in thicc blue columns, model predictions a solid red line.}
}
\description{
This function allows you to brush those time series without regrets
}
\examples{
model_painter(modell = sales_model) ## this will paint a nice observed vs. predicted time series plot
model_painter()
}
\keyword{all}
\keyword{and}
\keyword{dependent}
\keyword{from}
\keyword{happy}
\keyword{little}
\keyword{model}
\keyword{observations}
\keyword{paint}
\keyword{the}
\keyword{variable}
\keyword{your}
