% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_media.R
\name{plot_model_media}
\alias{plot_model_media}
\title{Quick draw of the media flightings used in the model}
\usage{
plot_model_media(modell = hans, startdatum = "2016-04-01")
}
\arguments{
\item{modell}{the model you have. Model predictions will be a solid red line.}

\item{startdatum}{first observation as date. The function will find the time period covered by the model by itself. Only accepts week right now.}
}
\description{
This function allows for a quick checking of plausibility towards the media channels by plotting them.
}
\examples{
plot_model_media(modell = lm_object, startdatum = "2018-01-01") ##
plot_model_media()
}
\keyword{flightings}
\keyword{media}
\keyword{paint}
