% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp}
\alias{pp}
\title{A function for fast printing of everything}
\usage{
pp(data, how_much = Inf)
}
\arguments{
\item{data}{Input data. Intended use for this function is in a tidyverse pipe, so you shouldn't have to specify this argument, usually}

\item{how_much}{How many lines you want printed. Helpful for when those 10 lines in a default tibble aren't enough....}
}
\description{
This function allows you a quick printout of all that's relevant to you
}
\examples{
pp(15) ## this will print 15 lines instead of the tidyverse standard of 10
pp()
}
\keyword{default}
\keyword{more}
\keyword{print}
\keyword{than}
\keyword{the}
