% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\title{A function for rapid sales decomposition}
\usage{
preview(fit = out)
}
\arguments{
\item{fit}{Input data. Just name the model you want to inspect}
}
\description{
This function allows you to quickly view the sales decomposition for your model.
NOTE: This function was orignally named preview() and was written entirely by Johannes Ritter, Senior Consultant at Annalect Hamburg
}
\examples{
preview(my_default_model) ## use the object assignment operator ' <- ' to include the dummy into your dataset
preview()
}
\keyword{decomposition}
\keyword{preview}
\keyword{quick}
\keyword{sales}
