% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_torte.R
\name{preview_torte}
\alias{preview_torte}
\title{A function for displaying the pie chart aggregation of the sales decomposition}
\usage{
preview_torte(modell, coler = "def", rad = 1.01, font_size = 0.9)
}
\arguments{
\item{modell}{Input data. Just name the model you want to inspect}

\item{coler}{color palette. Accepts a vector of length k+1, with k being the number of variables in the model. Defaults to the rainbow color sequence.}

\item{rad}{Controls the size of the pie. Defaults to slighty above 1.}

\item{font_size}{Size of the text in the chart. Defaults to 0.9.}
}
\description{
This function allows you to quickly view the sales decomposition for your model.
}
\examples{
preview_torte(my_default_model)
preview_torte()
}
\keyword{chart.}
\keyword{decomposition}
\keyword{pie}
\keyword{preview}
\keyword{quick}
\keyword{sales}
