% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rausschreiber.R
\name{rausschreiber}
\alias{rausschreiber}
\title{A function for collecting and exporting data for further excel makro-based processing}
\usage{
rausschreiber(modell = hans)
}
\arguments{
\item{modell}{the model you want to export data for. .lm files contain a reference to the original data frame whis i used here.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
rausschreiber()
}
\keyword{excel}
\keyword{export}
\keyword{to}
