% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_clutter}
\alias{remove_clutter}
\title{Utilities for the maRketingscience package}
\usage{
remove_clutter(input)
}
\description{
This function allows you to quickly create a dummy variable for controlling outliers
}
\examples{
write_quick_sd()
}
\keyword{for}
\keyword{maRketingscience}
\keyword{utilities}
