% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dummy.R
\name{set_dummy}
\alias{set_dummy}
\title{A function for rapid creation of an outlier catch dummy variable}
\usage{
set_dummy(model = model1, sd = 1)
}
\arguments{
\item{model}{Input data. It's just name the model that you want the dummy for.}

\item{sd}{Controls the cut-off value for the dummy variable that is being generated. Can be a non-integer like 1.5. Defaults to 1.}
}
\description{
This function allows you to quickly create a dummy variable for controlling outliers
}
\examples{
set_dummy(model1) ## use the object assignment operator ' <- ' to include the dummy into your dataset
set_dummy()
}
\keyword{create}
\keyword{dummy}
\keyword{variable}
