% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_quick_sd.R
\name{write_quick_sd}
\alias{write_quick_sd}
\title{A function for rapid creation of an outlier catch dummy variable}
\usage{
write_quick_sd(modell = hans, name = "quick_sd")
}
\arguments{
\item{modell}{Input data. Just name the model you want to create a dummy for}

\item{name}{any character string. will be concatenated with a .csv suffix, so you don't have to put it there}
}
\description{
This function allows you to quickly create a dummy variable for controlling outliers
}
\examples{
write_quick_sd(my_default_model, name="model_SD_01") ## will be written into the defined directory
write_quick_sd()
}
\keyword{SD}
\keyword{a}
\keyword{away}
\keyword{fast}
\keyword{write}
