% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dummy.R
\name{set_dummy}
\alias{set_dummy}
\title{A function for rapid creation of an outlier catch dummy variable}
\usage{
set_dummy(modell)
}
\arguments{
\item{modell}{Input data. Just name the model you want to create a dummy for}
}
\description{
This function allows you to quickly create a dummy variable for controlling outliers
}
\examples{
set_dummy(my_default_model) ## use the object assignment operator ' <- ' to include the dummy into your dataset
set_dummy()
}
\keyword{create}
\keyword{dummy}
\keyword{variable}
